// src/enums/socket/actions.ts
export enum Actions {
  JOIN = 'join',
  LEAVE = 'leave',
  SHARE_ROOMS = 'share-rooms',
  ADD_PEER = 'add-peer',
  REMOVE_PEER = 'remove-peer',
  RELAY_SDP = 'relay-sdp',
  RELAY_ICE = 'relay-ice',
  SESSION_DESCRIPTION = 'session-description',
  DISCONNECTING = 'disconnecting',
  ICE_CANDIDATE = 'ice-candidate',
}